refactor: Timer 분리
