+++ src/main/java/com/grepp/spring/app/controller/api/timer/TimerController.java
@@ -1,0 +1,44 @@
+package com.grepp.spring.app.controller.api.timer;
+
+import com.grepp.spring.app.model.timer.service.TimerService;
+import com.grepp.spring.infra.response.CommonResponse;
+import com.grepp.spring.infra.response.SuccessCode;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController("/api/v1/timer")
+@RequiredArgsConstructor
+public class TimerController {
+
+    private final TimerService timerService;
+
+    // 스터디별 공부시간 등록
+    @PostMapping("/{studyId}/{studyMemberId}")
+    @ApiResponse(responseCode = "200")
+    public ResponseEntity<CommonResponse<SuccessCode>> recordStudyTime(
+        @PathVariable Long studyId, @PathVariable Long studyMemberId
+    ) {
+       timerService.recordStudyTime(studyId, studyMemberId);
+       return ResponseEntity.ok(CommonResponse.success(SuccessCode.SUCCESS));
+    }
+
+    // 타이머 전체 누적 시간 조회
+    @GetMapping("/{memberId}/all-timer")
+    @ApiResponse(responseCode = "200")
+    public ResponseEntity<CommonResponse<Long>> getAllTimer(@PathVariable Long memberId) {
+        log.info("getAllTimer memberId: {}", memberId);
+        return ResponseEntity.ok(CommonResponse.success(timerService.getAllStudyTime(memberId)));
+    }
+
+    // 스터디별 7일단의 스터디별 공부시간 확인
+
+
+
+}
