+++ src/main/java/com/grepp/spring/app/model/timer/repository/TimerRepository.java
@@ -1,6 +1,6 @@
-package com.grepp.spring.app.model.member.repository;
+package com.grepp.spring.app.model.timer.repository;
 
-import com.grepp.spring.app.model.member.entity.Timer;
+import com.grepp.spring.app.model.timer.entity.Timer;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
